refactor export widgets
